Sideways triangle.
Refactored printRow.
