+++ main.go
@@ -1,7 +1,14 @@
 package main
 
-import "github.com/fonglh/think-like-a-programmer/chapter2"
+import (
+	"fmt"
+	"github.com/fonglh/think-like-a-programmer/chapter2"
+)
 
 func main() {
+	fmt.Println("Half Square")
 	chapter2.Halfsquare()
+
+	fmt.Println("Sideways Triangle")
+	chapter2.Sidewaystriangle()
 }
